pyproject.toml
[project]
name = "agent-backend"
version = "0.1.0"
description = "Universal AI tool-calling backend (enterprise architecture)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "memory-base>=0.1.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "watchdog>=4.0.0",
    "httpx>=0.26.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",
    "minio>=7.2.0",
    "pgvector>=0.2.0",
    "python-multipart>=0.0.6",
    "structlog>=24.1.0",
]

[project.scripts]
agent-backend = "app.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "aiosqlite>=0.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "real_ai: integration tests that call real AI APIs",
    "real_local: integration tests that run real local binaries (cursor agent, echo, date)",
]
